core: rename file to be consistent
