+++ generators/generator-base-docker.js
