+++ generators/openshift/index.js
@@ -20,7 +20,7 @@ const chalk = require('chalk');
 const shelljs = require('shelljs');
 const prompts = require('./prompts');
 const writeFiles = require('./files').writeFiles;
-const BaseDockerGenerator = require('../generator-docker-base');
+const BaseDockerGenerator = require('../generator-base-docker');
 const { loadFromYoRc, checkImages, generateJwtSecret, configureImageNames, setAppsFolderPaths } = require('../docker-base');
 const statistics = require('../statistics');
